Fix need to run tests with from_json_test.exs
Move assert_idempotent to `Alembic.FromJsonCase` in
`test/support/from_json_case.ex`, which is always available.
